fix if

+++ entrypoint.sh
@@ -96,6 +96,8 @@ terraform ${TF_VERB} ${TF_PLAN} ${TF_VARSFILE} ${TF_AUTOAPPROVE} ${TF_OUT}
 # Copy tfplan to github workspace
 if [ "$TF_VERB" = "plan" ]
 then
-    rm /github/workspace/$INPUT_PLANFILE
-    cp tfplan /github/workspace/$INPUT_PLANFILE
+    if [ ! -f /github/workspace/$INPUT_PLANFILE ]
+    then
+      cp tfplan /github/workspace/$INPUT_PLANFILE
+    fi
 fi
